refactor: 데이터베이스 관련 패키지 위치를 변경
